profiler: unit: --than ns, --order-mem B
--than, The default unit is ns, you can specify s, ms, us, ns. For example: 10ms

--order-mem, The default unit is bytes, you can specify GB, MB, KB, B, case-
insensitive. For example, 4MB

+++ monitor.h
@@ -64,7 +64,7 @@ struct env {
     char *filter;
     bool interruptible;
     bool uninterruptible;
-    int greater_than;
+    unsigned long greater_than; // unit: ns, percent
     bool callchain;
     int mmap_pages;
     bool exclude_user;
